+++ src/BigDecimal.php
@@ -335,14 +335,22 @@ final readonly class BigDecimal extends BigNumber
      * @param BigNumber|int|float|string $min The minimum. Must be convertible to a BigDecimal.
      * @param BigNumber|int|float|string $max The maximum. Must be convertible to a BigDecimal.
      *
-     * @throws MathException If min/max are not convertible to a BigDecimal.
+     * @throws MathException            If min/max are not convertible to a BigDecimal.
+     * @throws InvalidArgumentException If min is greater than max.
      */
     public function clamp(BigNumber|int|float|string $min, BigNumber|int|float|string $max): BigDecimal
     {
+        $min = BigDecimal::of($min);
+        $max = BigDecimal::of($max);
+
+        if ($min->isGreaterThan($max)) {
+            throw new InvalidArgumentException('Minimum value must be less than or equal to maximum value.');
+        }
+
         if ($this->isLessThan($min)) {
-            return BigDecimal::of($min);
+            return $min;
         } elseif ($this->isGreaterThan($max)) {
-            return BigDecimal::of($max);
+            return $max;
         }
 
         return $this;